requirements.txt
fastapi==0.115.6
uvicorn==0.27.1
pandas==2.2.1
python-dotenv==1.0.1
langchain==0.1.11
langchain-community==0.0.29
langchain-core==0.1.30
chromadb==0.4.24
sentence-transformers==2.5.1
langchain-groq==0.0.6
pydantic==2.6.3
python-multipart==0.0.9